Functionality update, code revision, credits update
Revision of the code with the help of Maestra Fénix and Obsidian Conflict Team. The map and strings lookup code was improved to load map name earlier and not have a limited number of strings.

The files were moved around and renamed a bit. Map-specific files now reside in vgui/loading/maps, and the loading wheel graphics in vgui/loading/ui, as the current code looks through vgui/loading directory to load any VMTs in it to use as generic/randomised backgrounds.